Revise VL_MaskMemToScreen

+++ id_vl.h
@@ -121,7 +121,7 @@ void VL_DrawPicBare (Sint16 x, Sint16 y, Uint8* pic, Sint16 width, Sint16 height
 void VL_MemToLatch(const Uint8* source, int width, int height, int dest);
 void VL_ScreenToScreen(int source, int dest, int width, int height);
 void VL_MemToScreen(const Uint8* source, int width, int height, int x, int y);
-void VL_MaskMemToScreen (const Uint8* source, int width, int height, int x, int y, int mask);
+void VL_MaskMemToScreen(const Uint8* source, int width, int height, int x, int y, Uint8 mask);
 void VL_ScreenToMem(Uint8* dest, int width, int height, int x, int y);
 
 void VL_DrawTile8String (char *str, char* tile8ptr, Sint16 printx, Sint16 printy);
